editPage % testPage js

+++ src/main/java/com/bionic/DAO/ResultDAO.java
@@ -70,11 +70,11 @@ public class ResultDAO extends AbstractDAO<Result> {
         return (BigInteger) query.getSingleResult();
     }
 
-    public List<Integer> getResultIdsByGroupAndTest(long testId, String groupName){
+    public List<BigInteger> getResultIdsByGroupAndTest(long testId, String groupName){
         Query query = em.createNamedQuery("getResultIdsByGroupAndTest");
         query.setParameter("testId", testId);
         query.setParameter("groupName", groupName);
-        return (List<Integer>)query.getResultList();
+        return (List<BigInteger>)query.getResultList();
     }
 
     public List<TestDTO> getAvailableTestsNames(long id, Permission permission) {
